Rephrase a log message in compile_dynamic_pcrs_job_list()
Devide the error code and its meaning with a colon.

Call the pcrs job dynamic and not the filter. Filters may
contain dynamic and non-dynamic pcrs jobs at the same time.
Only mention the name of the filter or tagger, but don't
claim it's a filter when it could be a tagger.