In server_content_encoding(), mark the content as taboo for filtering if SDCH compression is used
If SDCH was combined with a supported compression algorithm,
we'd previously try to decompress it, when successful apply
the enabled filters and ditch the Content-Encoding header
even though the SDCH compression wasn't removed.

Reported by zebul666 in #3225863.

I get the impression that filtering SDCH-compressed content
actually isn't guaranteed to cause problems as long as the
client still gets a "Content-Encoding: sdch" header, so we
should probably eventually allow users who want to risk it
to do so by explicitly enforcing filtering.

While at it, mark a bunch of other, partly-related, problems
in server_content_encoding() that should be addressed someday.