In bind_port(), stop hinting AI_ADDRCONFIG to getaddrinfo()
It's not strictly needed there and causes GNU libc to fail to
return loopback addresses if no non-loopback IP address is
configured on the system. Apparently this is the result of a
somewhat strict but valid interpretation of RFC 3493 and
considered a feature. Sad but true.

This is mainly an issue if the system is using DHCP and Privoxy
is started before the network is completely configured.

Reported by Raphael Marichez in #3349356.
Additional insight from Petr Pisar.