Redirect requests for http://config.privoxy.org/.* to http://www.privoxy.org/config/
config.privoxy.org used to be a separate server that was
administrated by Andreas, but will be relocated to the SF
webserver after the domain transfer is done.

Note that the things SF calls VHOSTS all point to the
same root directory, so we can't use a dedicated .htaccess
file for config.privoxy.org.