Remove an unreasonable parameter limit in parse_cgi_parameters()
The new parameter limit depends on the memory available and
is currently unlikely to be reachable, due to other limits in
both Privoxy and common clients.

So the limit is actually still there, but at least it's no
longer unreasonable, is properly handled, and should be high
enough for anybody.

Reported by Andrew on ijbswa-users@.