Add support for chunk-encoded client request bodies
It's simplistic and requires the whole request body to be buffered
before the request is forwarded. If the buffer limit is reached,
the request is rejected.

Previously chunk-encoded request bodies weren't guaranteed
to be forwarded correctly, so this can also be considered a
bug fix although chunk-encoded request bodies aren't commonly
used in the real world.