get_destination_from_headers(): Additionally update the request line in proxy format
This makes rewriting intercepted requests more convenient.

Previously it was expected to fail unless $hostport
was being used, but rewrites of intercepted requests
without $hostport failed "the wrong way" and would
result in an out-of-memory message (vanilla host patterns)
or a crash (extended host patterns).

Reported by "Guybrish Threepwood" in #1694.