Privoxy-Regression-Test: Add a "Default level offset" directive
... which can be used to change the default level by a given value.

This directive affects all tests located after it until the end
of the file or a another "Default level offset" directive is reached.

The purpose of this directive is to make it more convenient to skip
similar tests in a given file without having to remove or disable
the tests completely.