Revert r1.165 which didn't perform as advertised
While the idea was to use "https:// when creating links
for the user manual on the website", the actual effect
was to use "https://" when Privoxy was supposed to serve
the user manual itself.

Reverting the commit is sufficient as the "use https://
when creating links for the user manual on the website"
use case is already covered by project.h r1.214.

Reported by Yossi Zahn on Privoxy-devel@.