Remove a mysterious comment with a GNU FDL link
... as it isn't useful and could confuse license scanners.

In May 2002 it was briefly claimed that "this document" was covered
by the GNU FDL. The commit message (r1.5) doesn't explain the motivation
or whether all copyright holders were actually asked and agreed to the
declared license change.

It's thus hard to tell whether or not the license change was legit,
but luckily two days later the "doc license" was "put" "back to GPL"
anyway (r1.6).

At the same time the offending comment with a link to the FDL
(not the GPL) was added for no obvious reason.

Now it's gone again.